Fixed bottom navigation pushing screens issue
