+++ mobile-app/lib/screens/verification/verification_screen.dart
@@ -89,7 +89,7 @@ class _VerificationScreenState extends State<VerificationScreen> {
           }
           if (state is VerifySuccess) {
             Application.router
-                .navigateTo(context, Routes.home, clearStack: true);
+                .navigateTo(context, Routes.main, clearStack: true);
           }
         },
         cubit: Provider.of<VerifyBloc>(context),
